requirements.txt
langchain
openai
faiss-cpu
pdfplumber
tiktoken
streamlit
google-generativeai